requirements.txt
arrow
pinecone-client
PyPDF2
python-dotenv
tenacity
tiktoken
unstructured[all-docs]
xata
dashscope